remove unused view utils, move xml sitemap test to base_api
