+++ aleph/views/stream_api.py
@@ -1,12 +1,24 @@
 import logging
 from banal import ensure_list
-from flask import Blueprint, request
+from flask import Blueprint, request, abort
+from flask.wrappers import Response
 
 from aleph.index.entities import iter_entities, PROXY_INCLUDES
-from aleph.views.util import require, stream_ijson
+from aleph.util import JSONEncoder
 
 log = logging.getLogger(__name__)
 blueprint = Blueprint("bulk_api", __name__)
+
+def stream_ijson(iterable, encoder=JSONEncoder):
+    """Stream JSON line-based data."""
+
+    def _generate_stream():
+        for row in iterable:
+            row.pop("_index", None)
+            yield encoder().encode(row)
+            yield "\n"
+
+    return Response(_generate_stream(), mimetype="application/json+stream")
 
 
 @blueprint.route("/api/2/entities/_stream")
@@ -40,9 +52,11 @@ def entities(collection_id=None):
       - Entity
     """
     if collection_id is not None:
-        require(request.authz.can(collection_id, request.authz.WRITE))
+        if not request.authz.can(collection_id, request.authz.WRITE):
+            abort(403)
     else:
-        require(request.authz.is_admin)
+        if not request.authz.is_admin:
+            abort(403)
     schemata = ensure_list(request.args.getlist("schema"))
     includes = ensure_list(request.args.getlist("include"))
     includes = includes or PROXY_INCLUDES
